Don't die with python 3.4 if memprof is not yet installed
pkg_resources changed behaviour in python 3.4 and now raises
DistributionNotFound if the package you're building is not yet
installed.  When running the testsuite we import memprof in the current
directory.  However pkg_resources is unable to find it and we fail to
get the current version of memprof.  Set it to local in order to be able
to run the testsuite before installing the package.